+++ pages/user/admin/[id].vue
@@ -120,7 +120,7 @@ onBeforeRouteLeave(() => user.dispose());
 async function become() {
     await user.call('post', user.links?.self.href + '/become', undefined, {
         async onSuccess() {
-            const currentUser = useCurrentUserStore();
+            const currentUser = useCurrentUser();
             currentUser.fetch();
             await navigateTo('/');
         }