+++ pages/user/settings/apikeys.vue
@@ -97,7 +97,7 @@
 definePageMeta({
     middleware: ['auth']
 });
-const currentUser = useCurrentUserStore();
+const currentUser = useCurrentUser();
 const links = userSettingsLinks();
 const buttons = userSettingsButtons();
 const apikeys = useVingKind({