+++ pages/user/settings/index.vue
@@ -45,7 +45,7 @@
 definePageMeta({
     middleware: ['auth']
 });
-const currentUser = useCurrentUserStore();
+const currentUser = useCurrentUser();
 const links = userSettingsLinks();
 const buttons = userSettingsButtons();
 </script>